domain/key: change API and move Key struct impls to global/key.rs
